% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randall.R
\name{randall}
\alias{randall}
\title{randall function}
\usage{
randall(
  n,
  nmat,
  input,
  description,
  ord = c(1, 2, 3, 2, 1, 1, 2, 3, 2, 1, 2, 3, 1, 2, 1)
)
}
\arguments{
\item{n}{number of variables (rows in the matrix)}

\item{nmat}{number of matrices to be analyzed}

\item{input}{name of the input file}

\item{description}{any information you wish to enter describing each sample}

\item{ord}{prediction ordering (default 6 type RIASEC circumplex)}
}
\value{
Data frame of RTHOR model results, one row per matrix
}
\description{
Randomization test of hypothesized order relations
}
\examples{
randall_output <- RTHORR::randall(n=6,
                                  nmat=3,
                                  input=system.file("extdata", "input.txt", package = "RTHORR"),
                                  description = c("sample_one", "sample_two", "sample_three"))

}
