% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randall_from_df.R
\name{randall_from_df}
\alias{randall_from_df}
\title{randall function}
\usage{
randall_from_df(df_list, description, ord = "circular6")
}
\arguments{
\item{df_list}{list of dataframes to be run, dataframes input should have only columns for the variables included in the models (e.g., 6 columns for RIASEC)}

\item{description}{any information you wish to enter describing each sample}

\item{ord}{prediction ordering (default "circular6").
For circular models there are two preset inputs for 6 and 8 variables, "circular6" and "circular8".
Also accepts a vector of prediction ordering, for example circular6 = c(1,2,3,2,1,1,2,3,2,1,2,3,1,2,1)}
}
\value{
Data frame of RTHOR model results, one row per matrix
}
\description{
Randomization test of hypothesized order relations
}
\examples{
randall_output <- RTHORR::randall_from_df(df_list = list(data1, data2),
                                          ord = "circular6",
                                          description = c("sample_one", "sample_two", "sample_three"))

}
