% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randmf.R
\name{randmf}
\alias{randmf}
\title{randmf function}
\usage{
randmf(n, nmat, ord = c(1, 2, 3, 2, 1, 1, 2, 3, 2, 1, 2, 3, 1, 2, 1), input)
}
\arguments{
\item{n}{number of variables (rows in the matrix)}

\item{nmat}{number of matrices to be analyzed}

\item{ord}{prediction ordering (default 6 type RIASEC circumplex)}

\item{input}{name of the input file}
}
\value{
Data frame of RTHOR model results, one row per matrix
}
\description{
comparison of correlation matrices using randomization test of hypothesized order relations
    outputs two objects in a list: first the RTHOR results for each model, second the comparisons
}
\examples{
randmf_results <- randall(n=6,nmat=3,ord=c(1,2,3,2,1,1,2,3,2,1,2,3,1,2,1),input="h:/my documents/all/r/sifbend.txt")


}
