% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randmf_from_df.R
\name{randmf_from_df}
\alias{randmf_from_df}
\title{randmf_from_df function}
\usage{
randmf_from_df(df_list, ord = "circular6")
}
\arguments{
\item{df_list}{list of dataframes to be run, dataframes input should have only columns for the variables included in the models (e.g., 6 columns for RIASEC)}

\item{ord}{prediction ordering (default "circular6").
For circular models there are two preset inputs for 6 and 8 variables, "circular6" and "circular8".
Also accepts a vector of prediction ordering, for example circular6 = c(1,2,3,2,1,1,2,3,2,1,2,3,1,2,1)}
}
\value{
Data frame of RTHOR model results, one row per matrix
}
\description{
comparison of correlation matrices using randomization test of hypothesized order relations
    outputs two objects in a list: first the RTHOR results for each model, second the comparisons
}
\examples{
randmf_results <- RTHORR::randmf_from_df(df_list = list(data1, data2),
                                         ord="circular6")

}
